+++ app/src/main/java/dev/leonlatsch/photok/settings/ui/SettingsFragment.kt
@@ -39,6 +39,7 @@ import dev.leonlatsch.photok.databinding.BindingConverters
 import dev.leonlatsch.photok.other.extensions.show
 import dev.leonlatsch.photok.other.extensions.startActivityForResultAndIgnoreTimer
 import dev.leonlatsch.photok.other.setAppDesign
+import dev.leonlatsch.photok.other.statusBarPadding
 import dev.leonlatsch.photok.settings.data.Config
 import dev.leonlatsch.photok.settings.ui.changepassword.ChangePasswordDialog
 import dev.leonlatsch.photok.settings.ui.hideapp.ToggleAppVisibilityDialog
@@ -62,12 +63,7 @@ class SettingsFragment : PreferenceFragmentCompat() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        // Set top padding for layout
-        view.setOnApplyWindowInsetsListener { v, insets ->
-            v.setPadding(0, insets.top(), 0, 0)
-            insets
-        }
+        view.statusBarPadding()
 
         toolbar = view.findViewById(R.id.settingsToolbar)
     }
@@ -218,17 +214,3 @@ class SettingsFragment : PreferenceFragmentCompat() {
         const val KEY_ACTION_ABOUT = "action_about"
     }
 }
-
-/**
- * Thx mozilla
- *
- * https://github.com/mozilla-mobile/android-components/pull/9680/files#diff-9d900219329132b059f18f83b6e2952c5509bcfbf063a571ee5d647f76fa6554
- */
-fun WindowInsets.top(): Int =
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-        this.getInsets(WindowInsets.Type.systemBars()).top
-
-    } else {
-        @Suppress("DEPRECATION")
-        this.systemWindowInsetTop
-    }
